# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AncesTree",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/relationships": {
      "post": {
        "tags": [
          "relationships"
        ],
        "summary": "Create Relationship",
        "description": "Link two people. Refused with 409 and the reason when a rule would break.",
        "operationId": "create_relationship",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationshipResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/relationships/{link_id}": {
      "patch": {
        "tags": [
          "relationships"
        ],
        "summary": "Update Relationship",
        "operationId": "update_relationship",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationshipResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "relationships"
        ],
        "summary": "Delete Relationship",
        "operationId": "delete_relationship",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/relationship-kinds": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Relationship Kinds",
        "description": "Kinds of parent-child link, with how many links use each.",
        "operationId": "list_relationship_kinds",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KindView"
                  },
                  "type": "array",
                  "title": "Response List Relationship Kinds"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Create Relationship Kind",
        "operationId": "create_relationship_kind",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KindInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KindView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/relationship-kinds/{key}": {
      "patch": {
        "tags": [
          "settings"
        ],
        "summary": "Update Relationship Kind",
        "description": "Rename, relabel, hide (active=false), reorder, or change ring placement.",
        "operationId": "update_relationship_kind",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KindUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KindView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "settings"
        ],
        "summary": "Delete Relationship Kind",
        "description": "Only a kind that no link uses; hide it otherwise.",
        "operationId": "delete_relationship_kind",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenderedLabel": {
        "properties": {
          "neutral": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Neutral"
          },
          "male": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Male"
          },
          "female": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Female"
          }
        },
        "type": "object",
        "required": [
          "neutral"
        ],
        "title": "GenderedLabel",
        "description": "One side of a relationship kind: \"adoptive parent\", or \"adoptive father\"/\"mother\"."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KindInput": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Label"
          },
          "parent_label": {
            "$ref": "#/components/schemas/GenderedLabel"
          },
          "child_label": {
            "$ref": "#/components/schemas/GenderedLabel"
          },
          "in_layout": {
            "type": "boolean",
            "title": "In Layout",
            "default": true
          },
          "words": {
            "additionalProperties": {
              "$ref": "#/components/schemas/KindWords"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/KindLanguage"
            },
            "type": "object",
            "title": "Words"
          }
        },
        "type": "object",
        "required": [
          "label",
          "parent_label",
          "child_label"
        ],
        "title": "KindInput"
      },
      "KindLanguage": {
        "type": "string",
        "enum": [
          "ms",
          "jv"
        ]
      },
      "KindUpdate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "parent_label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenderedLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "child_label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenderedLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "in_layout": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Layout"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          },
          "words": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/KindWords"
                },
                "propertyNames": {
                  "$ref": "#/components/schemas/KindLanguage"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Words"
          }
        },
        "type": "object",
        "title": "KindUpdate"
      },
      "KindView": {
        "properties": {
          "key": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "builtin": {
            "type": "boolean",
            "title": "Builtin"
          },
          "blood": {
            "type": "boolean",
            "title": "Blood"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "in_layout": {
            "type": "boolean",
            "title": "In Layout"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "parent_label": {
            "$ref": "#/components/schemas/GenderedLabel"
          },
          "child_label": {
            "$ref": "#/components/schemas/GenderedLabel"
          },
          "words": {
            "additionalProperties": {
              "$ref": "#/components/schemas/KindWords"
            },
            "type": "object",
            "title": "Words"
          },
          "usage": {
            "type": "integer",
            "title": "Usage"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "builtin",
          "blood",
          "active",
          "in_layout",
          "sort_order",
          "parent_label",
          "child_label",
          "usage"
        ],
        "title": "KindView"
      },
      "KindWords": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Label"
          },
          "parent_label": {
            "$ref": "#/components/schemas/GenderedLabel"
          },
          "child_label": {
            "$ref": "#/components/schemas/GenderedLabel"
          }
        },
        "type": "object",
        "required": [
          "label",
          "parent_label",
          "child_label"
        ],
        "title": "KindWords",
        "description": "A kind's words in another kinship language, e.g. Malay: \"angkat\", \"bapa\nangkat\" / \"emak angkat\", \"anak angkat\". `label` is the kind as a qualifier, as in \"abang\nangkat\"."
      },
...
      "LinkView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "parent",
              "spouse"
            ],
            "title": "Type"
          },
          "source": {
            "type": "string",
            "format": "uuid",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "format": "uuid",
            "title": "Target"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpouseStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "source",
          "target",
          "kind",
          "status"
        ],
        "title": "LinkView"
      },
...
      "Notice": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Notice",
        "description": "Something worth a second look that was still saved, e.g. an unlikely age gap."
      },
...
      "RelationshipCreate": {
        "properties": {
          "person_a": {
            "type": "string",
            "format": "uuid",
            "title": "Person A"
          },
          "person_b": {
            "type": "string",
            "format": "uuid",
            "title": "Person B"
          },
          "a_is": {
            "type": "string",
            "enum": [
              "parent",
              "child",
              "spouse",
              "sibling"
            ],
            "title": "A Is"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "biological"
          },
          "status": {
            "$ref": "#/components/schemas/SpouseStatus",
            "default": "married"
          },
          "shared_parents": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shared Parents"
          }
        },
        "type": "object",
        "required": [
          "person_a",
          "person_b",
          "a_is"
        ],
        "title": "RelationshipCreate",
        "description": "Read as \"`person_a` is `person_b`'s `a_is`\": a is b's mother -> a_is \"parent\"."
      },
      "RelationshipResult": {
        "properties": {
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkView"
            },
            "type": "array",
            "title": "Links"
          },
          "notices": {
            "items": {
              "$ref": "#/components/schemas/Notice"
            },
            "type": "array",
            "title": "Notices"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/Suggestion"
            },
            "type": "array",
            "title": "Suggestions"
          }
        },
        "type": "object",
        "required": [
          "links",
          "notices",
          "suggestions"
        ],
        "title": "RelationshipResult"
      },
      "RelationshipUpdate": {
        "properties": {
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpouseStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "swap": {
            "type": "boolean",
            "title": "Swap",
            "default": false
          }
        },
        "type": "object",
        "title": "RelationshipUpdate"
      },
...
      "SpouseStatus": {
        "type": "string",
        "enum": [
          "married",
          "divorced",
          "widowed"
        ],
        "title": "SpouseStatus"
      },
      "Suggestion": {
        "properties": {
          "type": {
            "type": "string",
            "const": "marry",
            "title": "Type"
          },
          "person_a": {
            "type": "string",
            "format": "uuid",
            "title": "Person A"
          },
          "person_b": {
            "type": "string",
            "format": "uuid",
            "title": "Person B"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "type",
          "person_a",
          "person_b",
          "message"
        ],
        "title": "Suggestion"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}